Moved some code around to clean it up a bit
